+++ airflow/bin/cli.py
@@ -1633,8 +1633,14 @@ def list_dag_runs(args, dag=None):
 @cli_utils.action_logging
 def sync_perm(args):
     appbuilder = cached_appbuilder()
-    print('Update permission, view-menu for all existing roles')
+    print('Updating permission, view-menu for all existing roles')
     appbuilder.sm.sync_roles()
+    print('Updating permission on all DAG views')
+    dags = DagBag().dags.values()
+    for dag in dags:
+        appbuilder.sm.sync_perm_for_dag(
+            dag.dag_id,
+            dag.access_control)
 
 
 Arg = namedtuple(
@@ -2293,7 +2299,7 @@ class CLIFactory(object):
         },
         {
             'func': sync_perm,
-            'help': "Update existing role's permissions.",
+            'help': "Update permissions for existing roles and DAGs.",
             'args': tuple(),
         },
         {
